pyproject.toml
[tool.poetry]
name = "devops-course-starter"
version = "0.1.0"
description = ""
authors = ["Oskar Williams <oskarwilliams378@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
atomicwrites = "^1.4.0"
attrs = "^19.3.0"
certifi = "^2020.6.20"
chardet = "^3.0.4"
click = "^7.1.2"
colorama = "^0.4.3"
dominate = "^2.5.1"
Flask = "^1.1.1"
gunicorn = "^20.0.4"
idna = "^2.9"
itsdangerous = "^1.1.0"
Jinja2 = "^2.11.2"
MarkupSafe = "^1.1.1"
more-itertools = "^8.4.0"
packaging = "^20.4"
pluggy = "^0.13.1"
py = "^1.9.0"
pyparsing = "^2.4.7"
pytest = "^5.4.3"
python-dateutil = "^2.8.1"
python-dotenv = "^0.12.0"
requests = "^2.24.0"
selenium = "^3.141.0"
six = "^1.15.0"
urllib3 = "^1.25.9"
visitor = "^0.1.3"
wcwidth = "^0.2.5"
Werkzeug = "^1.0.1"
pymongo = {extras = ["srv"], version = "^3.11.2"}
mongomock = "^3.22.1"
oauthlib = "^3.1.0"
flask-login = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[virtualenvs]
create = true
in-project = false
